파일 이름 명료화 & 버전 별 설명 추가
